requirements.txt
asgiref==3.5.0
Django==4.0.1
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-extra-fields==3.2.1
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
